Add issueClose to own file Create initial push event handler
