+++ package.json
@@ -25,6 +25,7 @@
     "probot": "^9.5.3"
   },
   "devDependencies": {
+    "fs": "0.0.1-security",
     "jest": "^24.9.0",
     "jest-expect-message": "^1.0.2",
     "nock": "^12.0.0",
